Fixed bug in length of PacketMapChun (0x33)
Thanks Herman!

+++ src/main/java/ch/spacebase/mcprotocol/standard/packet/PacketMapChunk.java
@@ -85,7 +85,7 @@ public class PacketMapChunk extends Packet {
 		out.writeBoolean(this.groundUp);
 		out.writeShort((short) (this.startY & 0xffff));
 		out.writeShort((short) (this.endY & 0xffff));
-		out.writeByte(this.length);
+		out.writeInt(this.length);
 		out.write(this.data, 0, this.length);
 	}
